+++ src/main/java/duke/command/AddCommand.java
@@ -1,0 +1,9 @@
+package command;
+
+public class AddCommand extends Command{
+
+    @Override
+    public void run() {
+
+    }
+}
