+++ src/main/java/duke/command/AddEventCommand.java
@@ -1,0 +1,2 @@
+package duke.command;public class AddEventCommand {
+}
