+++ src/main/java/duke/command/AddTodoCommand.java
@@ -1,0 +1,2 @@
+package duke.command;public class AddTodoCommand {
+}
