+++ src/main/java/duke/command/Command.java
@@ -1,0 +1,15 @@
+package command;
+
+public abstract class Command {
+    private boolean isExit;
+
+    public Command() {
+        this.isExit = false;
+    }
+
+    public boolean isExit() {
+        return this.isExit;
+    }
+
+    public abstract void run();
+}
