+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,0 +1,9 @@
+package command;
+
+public class DeleteCommand extends Command {
+
+    @Override
+    public void run() {
+
+    }
+}
