+++ src/main/java/duke/command/ExitCommand.java
@@ -1,0 +1,10 @@
+package command;
+
+public class ExitCommand extends Command {
+
+
+    @Override
+    public void run() {
+        
+    }
+}
