+++ src/main/java/duke/command/ListCommand.java
@@ -1,0 +1,9 @@
+package command;
+
+public class ListCommand extends Command {
+
+    @Override
+    public void run() {
+
+    }
+}
