+++ src/main/java/duke/command/MarkCommand.java
@@ -1,0 +1,9 @@
+package command;
+
+public class MarkCommand extends Command {
+
+    @Override
+    public void run() {
+
+    }
+}
