+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,0 +1,9 @@
+package command;
+
+public class UnmarkCommand extends Command {
+
+    @Override
+    public void run() {
+
+    }
+}
